requirements.txt
Django==3.2.18
django-confy==1.0.4
whitenoise==5.3.0
git+https://github.com/dbca-wa/dbca-utils.git#egg=dbca-utils
aws-requests-auth==0.4.3
gunicorn==19.9.0
gevent==22.10.2
django-encrypted-model-fields==0.6.5
ipaddress==1.0.23
psycopg2-binary==2.8.5
pycryptodome=3.17
pycryptodomex=3.17
#pycrypto==2.6.1
qrcode==7.4.2